SimpleMazeFactoryを作成した
