+++ Codes/AbstractFactory/src/abstract_site.py
@@ -1,7 +1,7 @@
-class Wall:
+class MapSite:
 
     def __init__(self):
         ...
 
     def enter(self):
-        return "引き返す"
+        ...
